Fix loading of Google Drive API so that we only load it once (#215)
[#178634302]

In working through the TypeScript conversion I noticed that the GoogleDriveProvider was loading an instance of the Google Drive API in every function call rather than caching the result of the first load and returning it for all subsequent loads. This PR makes sure we only load the Google Drive API once and also includes a jest test which in good TDD fashion was written before the fix was implemented.