Added save hash check and fixed incorrect saved flag setting on file loads [#180438651]
This adds a early exit on saves if the hash of the content doesn't change between the first load and save or between commanded saves.

This also adds a test provider so that we can use jest tests to test the new code that was added.

The "saved" flag was being set on the load event instead of the dirty flag.  The update is to set the dirty flag to true if the content requires conversion.