Configure download in .rc file
CODAP has permitted configuration of Po interactions through a .rc file for some time.
This avoids having to specify access key on the command line.
The .rc is in the user's home directory and not in the codeline.
This fix adds support of the .rc file in the CFM.
Configuration through command line parameters remains possible.