Fix issue with CODAP file dialog buttons [#184137566]
This adds file dialog specific button styling to fix an issue with CODAP rendering the buttons outside the dialog due to its custom global navbar button css.

The issue is caused by the new filterable file listings.  As the height changes on the file list the buttons were not adjusting.