add share button to no-frame example
This fixes a bug when sharing plain text content.
It also fixes a bug that was breaking all sharing in these examples because the sharing provider uses at `process.env`.

The `process.env` reference still exists in the library, so any application using the CFM library needs to handle this reference on its own.